Droits pages

+++ src/main/java/fr/uha/miage/vod/controller/ActeurController.java
@@ -1,5 +1,6 @@
 package fr.uha.miage.vod.controller;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,11 +22,21 @@ public class ActeurController {
 
 	// Affiche le formulaire de création d'un acteur
 	@GetMapping("/acteurcreer")
-	public String acteurcreerform(Model model) {
-		model.addAttribute("acteur", new Acteur());
-		Iterable<Acteur> liste = acteurRepository.findAll();
-		model.addAttribute("acteurs", liste);
-		return "acteurcreer";
+	public String acteurcreerform(Model model, HttpSession session) {
+		
+		if ((session.getAttribute("id") != null) && ((int) session.getAttribute("role") == 1)) {
+			model.addAttribute("acteur", new Acteur());
+			Iterable<Acteur> liste = acteurRepository.findAll();
+			model.addAttribute("acteurs", liste);
+			return "acteurcreer";
+		}
+		else if(session.getAttribute("id") != null) {
+			return "redirect:/";
+		}
+		
+		else {
+			return "redirect:/utilisateurconnecter";
+		}
 	}
 
 	// Enregistre l'acteur créé, en verifiant qu'il corresponde aux critères
@@ -35,19 +46,27 @@ public class ActeurController {
 			return "acteurcreer";
 
 		acteurRepository.save(acteur);
-		
 
 		return "redirect:/acteur";
 	}
 
 	// Affiche le formulaire d'édition d'un acteur
 	@GetMapping("/acteurmodifier/{id}")
-	public String acteurmodifierform(@PathVariable("id") Long id, Model model) {
-		Acteur acteur = acteurRepository.findOne(id);
-		model.addAttribute("acteur", acteur);
-		Iterable<Acteur> liste = acteurRepository.findAll();
-		model.addAttribute("acteurs", liste);
-		return "acteurmodifier";
+	public String acteurmodifierform(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if ((session.getAttribute("id") != null) && ((int) session.getAttribute("role") == 1)) {
+			Acteur acteur = acteurRepository.findOne(id);
+			model.addAttribute("acteur", acteur);
+			Iterable<Acteur> liste = acteurRepository.findAll();
+			model.addAttribute("acteurs", liste);
+			return "acteurmodifier";
+		}
+		else if(session.getAttribute("id") != null) {
+			return "redirect:/";
+		}
+		
+		else {
+			return "redirect:/utilisateurconnecter";
+		}
 	}
 
 	// Enregistre l'acteur modifié, en vérifiant qu'il corresponde aux critères
@@ -62,16 +81,34 @@ public class ActeurController {
 
 	// Affiche la liste des acteurs
 	@GetMapping("/acteur")
-	public String acteur(Model model) {
-		Iterable<Acteur> liste = acteurRepository.findAll();
-		model.addAttribute("acteurs", liste);
-		return "acteur";
+	public String acteur(Model model, HttpSession session) {
+		if ((session.getAttribute("id") != null) && ((int) session.getAttribute("role") == 1)) {
+			Iterable<Acteur> liste = acteurRepository.findAll();
+			model.addAttribute("acteurs", liste);
+			return "acteur";
+		}
+		else if(session.getAttribute("id") != null) {
+			return "redirect:/";
+		}
+		
+		else {
+			return "redirect:/utilisateurconnecter";
+		}
 	}
 
 	// Supprime l'acteur sélectionné
 	@GetMapping("/acteursupprimer/{id}")
-	public String acteursupprimer(@PathVariable("id") Long id) {
-		acteurRepository.delete(id);
-		return "redirect:/acteur";
+	public String acteursupprimer(@PathVariable("id") Long id, HttpSession session) {
+		if ((session.getAttribute("id") != null) && ((int) session.getAttribute("role") == 1)) {
+			acteurRepository.delete(id);
+			return "redirect:/acteur";
+		}
+		else if(session.getAttribute("id") != null) {
+			return "redirect:/";
+		}
+		
+		else {
+			return "redirect:/utilisateurconnecter";
+		}
 	}
 }
